style, IRQL_t однобайтовый

+++ inc/paramdepend.h
@@ -808,7 +808,7 @@ extern "C" {
 
 #elif (__GIC_PRESENT == 1)
 
-	typedef uint32_t IRQL_t;
+	typedef uint8_t IRQL_t;
 
 	enum
 	{
